Expand ~ in config file names.
People often invoke bde_verify with --config=~/my.cfg and do not
realize that (at least some) shells fail to expand the ~.  Because
bde_verify tries to load local .bdeverify files, and for other similar
reasons, we do not want to issue a warning for a config file we can't
open, and so people wound up running with no configuration at all,
resulting in spew.

So now we look for a leading ~/ in file names, and if we find it, we
replace the ~ with $HOME and let the environment variable expansion
code do the right thing.