Require assignment for RV01 check.
To avoid false positives on the RV01 check (prefer returning by value
instead of through a pointer), emit the warning only when the function
definition contains a simple assignment through the pointer.

This means that the warning will no longer be emitted on declarations,
just on definitions.

Also, this restriction isn't as tight as it could be.  The full check
should be that the pointee isn't used in other ways, but this will do
for now.