Add -diff plugin arg for git diff restrictions.
Specifying -diff=file (- for standard input) passes the arguments along to the
binary as '-Xclang -plugin-arg-bde_verify -Xclang diff=file'.  The file is read
as the output of a 'git diff command', in which lines beginning with '+++' are
used to accumulate filenames (with paths stripped) and subsequent lines
beginning with '@' are analyzed for '+LINE_NUMBER' or '+LINE_NUMBER,LINE_COUNT'
content.  Output warnings will be restricted to only the specified line numbers
of the specified files.