Fix calculation of whether constructors take allocators.
Normally, a constructor takes an allocator if its last parameter is an
allocator.  But if a class itself inherits from an allocator, its copy
constructor would look liek an allocator by this criterion.  So for copy
constructors, a parameter is only the last one if it's also not the only one.